twitter-api/requirements.txt
python-twitter
tweepy
requests
oauthlib